pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "air-bot"
version = "0.1.0"
description = "Maximum-Detail & Maximum-Efficiency Blueprint – detectors, scoring, async queue, DB, Kafka/Webhook"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Air-bot Team" }]
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.19",
    "asyncpg>=0.29",
    "aiokafka>=0.10",
    "httpx>=0.25",
    "fastapi>=0.109",
    "uvicorn[standard]>=0.24",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
]

[project.scripts]
airbot = "airbot.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["airbot*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.coverage.run]
source = ["airbot", "config", "db", "detectors", "scoring", "output", "collector", "soc_mint"]
omit = ["*/tests/*"]
